+++ Samples/Assets/GameCode/Samples.Dungeon.First/OuterWallTileComponent.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Entities;
+
+namespace Samples.Dungeon.First
+{
+    public struct OuterWallTile : IComponentData
+    {
+    }
+
+    public class OuterWallTileComponent : ComponentDataWrapper<OuterWallTile> { }
+}