+++ Samples/Assets/GameCode/Samples.Dungeon/BoardReferenceComponent.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using Unity.Entities;
+using UnityEngine;
+
+namespace Samples.Dungeon
+{
+    [Serializable]
+    public struct BoardReference : ISharedComponentData
+    {
+        public int TileSetId;
+    }
+    public class BoardReferenceComponent : SharedComponentDataWrapper<BoardReference> { }
+}