+++ Samples/Assets/GameCode/Samples.Dungeon/FloorTileComponent.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Entities;
+
+namespace Samples.Dungeon
+{
+    public struct FloorTile : IComponentData
+    {
+    }
+
+    public class FloorTileComponent : ComponentDataWrapper<FloorTile> { }
+}