+++ Samples/Assets/GameCode/Samples.Dungeon/WallTileComponent.cs
@@ -1,0 +1,10 @@
+﻿using Unity.Entities;
+
+namespace Samples.Dungeon
+{
+    public struct WallTile : IComponentData
+    {
+    }
+
+    public class WallTileComponent : ComponentDataWrapper<WallTile> { }
+}