+++ TwoStickShooter/Classic/Assets/GameCode/Components/Health.cs
@@ -1,9 +1,22 @@
-﻿using UnityEngine;
+﻿using System.Collections.Generic;
+using UnityEngine;
 
 namespace TwoStickClassicExample
 {
     public class Health : MonoBehaviour
     {
+        public static ICollection<Health> All => allHealths;
+        private static readonly HashSet<Health> allHealths = new HashSet<Health>();
+
+        private void Awake()
+        {
+            allHealths.Add(this);
+        }
+        
+        private void OnDestroy()
+        {
+            allHealths.Remove(this);
+        }
 
         public float Value
         {