+++ ReleaseNotes.md
@@ -1,3 +1,8 @@
+# 0.0.5
+## Changes
+* Throw ArgumentException when creating an entity with component data exceeding chunk size (64kb)
+* EntityManager.CreateComponentGroup is no longer public, use ComponentSystem.GetComponentGroup instead
+
 # 0.0.4
 
 ## New Features
@@ -34,9 +39,7 @@
 * UnityPackageManager -> Packages folder. (Unity 2018.1 beta 7 introduces this change and we reflected it in the sample project)
 
 * EntityManager.CreateComponentGroup should be replaced with ComponentSystem.GetComponentGroup.
-It automatically associates & caches the ComponentGroup with the system (It is automatically disposed by ComponentSystem) and thus input dependencies will be setup correctly.
-
-Additionally ComponentSystem.GetComponentGroup should not be called in OnUpdate() (Create and cache in OnCreateManager instead). ComponentSystem.GetComponentGroup allocates GC memory because the input is a param ComponentType[]...
+It automatically associates & caches the ComponentGroup with the system (It is automatically disposed by ComponentSystem) and thus input dependencies will be setup correctly. Additionally ComponentSystem.GetComponentGroup should not be called in OnUpdate() (It is recommended to create and cache in OnCreateManager instead). ComponentSystem.GetComponentGroup allocates GC memory because the input is a param ComponentType[]...
 
 * Systems are automatically disabled when all ComponentGroups have zero entities.
 [AlwaysUpdateSystem] can be used to always force update a system.
@@ -55,6 +58,8 @@ Additionally ComponentSystem.GetComponentGroup should not be called in OnUpdate(
 * Can now get an array of shared component data from a component group (ComponentGroup.GetSharedComponentDataArray)
   SharedComponentDataArray<T> can also be injected similar to ComponentDataArray<T>
   Access through SharedComponentDataArray is always read only
+
+* IJobProcessComponentData is significantly simplified. Supports 1, 2, 3 parameters. Supports read only, supports additional required components & subtractive components. https://github.com/Unity-Technologies/ECSJobDemos/blob/stable/ECSJobDemos/Assets/GameCode/SimpleRotation/RotationSpeedSystem.cs
 
 # 0.0.2